picstyle rework part 3
