+++ src/picstyle.c
@@ -4,6 +4,8 @@
 #include <bmp.h>
 #include <config.h>
 #include <picstyle.h>
+#include <lvinfo.h>
+#include <raw.h>
 
 // Converts picstyle ID to Canon menu position (index)
 picstyle_id get_picstyle_menu_id(picstyle_menu_index index)
@@ -280,3 +282,353 @@ LENS_SET_IN_PICSTYLE(contrast, -4, 4)
 LENS_SET_IN_PICSTYLE(sharpness, -1, 7)
 LENS_SET_IN_PICSTYLE(saturation, -4, 4)
 LENS_SET_IN_PICSTYLE(color_tone, -4, 4)
+
+
+#ifdef FEATURE_PICSTYLE
+
+// moved from lens.c
+static LVINFO_UPDATE_FUNC(picstyle_update)
+{
+    LVINFO_BUFFER(12);
+
+    if (is_movie_mode())
+    {
+        /* picture style has no effect on raw video => don't display */
+        if (raw_lv_is_enabled())
+            return;
+    }
+    else
+    {
+        /* when shooting RAW photos, picture style only affects the preview => don't display */
+        int jpg = pic_quality & 0x10000;
+        if (!jpg)
+            return;
+    }
+
+    snprintf(buffer, sizeof(buffer), "%s",
+        (char*)get_picstyle_name(lens_info.raw_picstyle)
+    );
+}
+
+// moved from shoot.c
+
+static void
+contrast_toggle( void * priv, int sign )
+{
+    int c = lens_get_contrast();
+    if (c < -4 || c > 4) return;
+    int newc = MOD((c + 4 + sign), 9) - 4;
+    lens_set_contrast(newc);
+}
+
+
+static MENU_UPDATE_FUNC(contrast_display)
+{
+    int s = lens_get_contrast();
+    MENU_SET_VALUE(
+        "%d",
+        s
+    );
+    MENU_SET_ICON(MNI_PERCENT, (s+4) * 100 / 8);
+}
+
+static void
+sharpness_toggle( void * priv, int sign )
+{
+    int c = lens_get_sharpness();
+    if (c < 0 || c > 7) return;
+    int newc = MOD(c + sign, 8);
+    lens_set_sharpness(newc);
+}
+
+static MENU_UPDATE_FUNC(sharpness_display)
+{
+    int s = lens_get_sharpness();
+    MENU_SET_VALUE(
+        "%d ",
+        s
+    );
+    MENU_SET_ICON(MNI_PERCENT, s * 100 / 7);
+}
+
+static void
+saturation_toggle( void * priv, int sign )
+{
+    int c = lens_get_saturation();
+    if (c < -4 || c > 4) return;
+    int newc = MOD((c + 4 + sign), 9) - 4;
+    lens_set_saturation(newc);
+}
+
+static MENU_UPDATE_FUNC(saturation_display)
+{
+    int s = lens_get_saturation();
+    int ok = (s >= -4 && s <= 4);
+    MENU_SET_VALUE(
+        ok ? 
+            "%d " :
+            "N/A",
+        s
+    );
+    MENU_SET_ENABLED(ok);
+    if (ok) MENU_SET_ICON(MNI_PERCENT, (s+4) * 100 / 8);
+    else { MENU_SET_ICON(MNI_OFF, 0); MENU_SET_WARNING(MENU_WARN_NOT_WORKING, "N/A"); }
+}
+
+static void
+color_tone_toggle( void * priv, int sign )
+{
+    int c = lens_get_color_tone();
+    if (c < -4 || c > 4) return;
+    int newc = MOD((c + 4 + sign), 9) - 4;
+    lens_set_color_tone(newc);
+}
+
+static MENU_UPDATE_FUNC(color_tone_display)
+{
+    int s = lens_get_color_tone();
+    int ok = (s >= -4 && s <= 4);
+    MENU_SET_VALUE(
+        ok ?
+            "%d " :
+            "N/A",
+        s
+    );
+    MENU_SET_ENABLED(ok);
+    if (ok) MENU_SET_ICON(MNI_PERCENT, (s+4) * 100 / 8);
+    else { MENU_SET_ICON(MNI_OFF, 0); MENU_SET_WARNING(MENU_WARN_NOT_WORKING, "N/A"); }
+}
+
+static CONFIG_INT("picstyle.rec", picstyle_rec, 0);
+static int picstyle_before_rec = 0; // if you use a custom picstyle during REC, the old one will be saved here
+
+static MENU_UPDATE_FUNC(picstyle_display)
+{
+    int i = picstyle_rec && RECORDING ? picstyle_before_rec : (int)lens_info.picstyle;
+
+    MENU_SET_VALUE(
+        get_picstyle_name(get_picstyle_menu_id(i))
+    );
+
+
+    if (picstyle_rec && is_movie_mode())
+    {
+        MENU_SET_RINFO(
+            "REC:%s",
+            get_picstyle_name(get_picstyle_menu_id(picstyle_rec))
+        );
+    }
+    else MENU_SET_RINFO(
+            "%d,%d,%d,%d",
+            lens_get_from_other_picstyle_sharpness(i),
+            lens_get_from_other_picstyle_contrast(i),
+            ABS(lens_get_from_other_picstyle_saturation(i)) < 10 ? lens_get_from_other_picstyle_saturation(i) : 0,
+            ABS(lens_get_from_other_picstyle_color_tone(i)) < 10 ? lens_get_from_other_picstyle_color_tone(i) : 0
+        );
+
+    MENU_SET_ENABLED(1);
+}
+
+static MENU_UPDATE_FUNC(picstyle_display_submenu)
+{
+    int p = get_picstyle_menu_id(lens_info.picstyle);
+    MENU_SET_VALUE(
+        "%s",
+        get_picstyle_name(p)
+    );
+    MENU_SET_ENABLED(1);
+}
+
+static void
+picstyle_toggle(void* priv, int sign )
+{
+    if (RECORDING) return;
+    int p = lens_info.picstyle;
+    p = MOD(p + sign - 1, NUM_PICSTYLES) + 1;
+    if (p)
+    {
+        p = get_picstyle_menu_id(p);
+        prop_request_change(PROP_PICTURE_STYLE, &p, 4);
+    }
+}
+
+#ifdef FEATURE_REC_PICSTYLE
+
+static MENU_UPDATE_FUNC(picstyle_rec_sub_display)
+{
+    if (!picstyle_rec)
+    {
+        MENU_SET_VALUE("OFF");
+        return;
+    }
+
+    MENU_SET_VALUE(
+        get_picstyle_name(get_picstyle_menu_id(picstyle_rec))
+    );
+    //~ MENU_SET_RINFO(
+    if (info->can_custom_draw) bmp_printf(MENU_FONT_GRAY, info->x_val, info->y + font_large.height,
+        "%d,%d,%d,%d",
+        lens_get_from_other_picstyle_sharpness(picstyle_rec),
+        lens_get_from_other_picstyle_contrast(picstyle_rec),
+        ABS(lens_get_from_other_picstyle_saturation(picstyle_rec)) < 10 ? lens_get_from_other_picstyle_saturation(picstyle_rec) : 0,
+        ABS(lens_get_from_other_picstyle_color_tone(picstyle_rec)) < 10 ? lens_get_from_other_picstyle_color_tone(picstyle_rec) : 0
+    );
+}
+
+static void
+picstyle_rec_sub_toggle( void * priv, int delta )
+{
+    if (RECORDING) return;
+    picstyle_rec = MOD(picstyle_rec+ delta, NUM_PICSTYLES+1);
+}
+
+static void rec_picstyle_change(int rec)
+{
+    static int prev = 0;
+
+    if (picstyle_rec)
+    {
+        if (prev == 0 && rec) // will start recording
+        {
+            picstyle_before_rec = lens_info.picstyle;
+            int p = get_prop_picstyle_from_index(picstyle_rec);
+            if (p)
+            {
+                NotifyBox(2000, "Picture Style : %s", get_picstyle_name(p));
+                prop_request_change(PROP_PICTURE_STYLE, &p, 4);
+            }
+        }
+        else if (prev == 2 && rec == 0) // recording => will stop
+        {
+            int p = get_prop_picstyle_from_index(picstyle_before_rec);
+            if (p)
+            {
+                NotifyBox(2000, "Picture Style : %s", get_picstyle_name(p));
+                prop_request_change(PROP_PICTURE_STYLE, &p, 4);
+            }
+            picstyle_before_rec = 0;
+        }
+    }
+    prev = rec;
+}
+
+#endif // FEATURE_REC_PICSTYLE
+
+static struct menu_entry picstyle_features_menu[] = {
+    {
+        .name = "Picture Style",
+        .update     = picstyle_display,
+        .select     = picstyle_toggle,
+        .priv = &lens_info.picstyle,
+        .help = "Change current picture style.",
+        .edit_mode = EM_SHOW_LIVEVIEW,
+        .icon_type = IT_DICE,
+        .choices = (const char *[]) {
+                #if NUM_PICSTYLES > 9 // 600D, 5D3...
+                "Auto",
+                #endif
+                "Standard", "Portrait", "Landscape",
+                #if NUM_PICSTYLES == 11 // D8 and up
+                "FineDetail",
+                #endif
+                "Neutral", "Faithful", "Monochrome", "UserDef1", "UserDef2", "UserDef3" },
+        .min = 1,
+        .max = NUM_PICSTYLES,
+        .submenu_width = 550,
+        .submenu_height = 300,
+        //~ .show_liveview = 1,
+        //~ //.essential = FOR_PHOTO | FOR_MOVIE,
+        .children =  (struct menu_entry[]) {
+            {
+                .name = "Picture Style",
+                .priv = &lens_info.picstyle,
+                .min = 1,
+                .max = NUM_PICSTYLES,
+                .choices = (const char *[]) {
+                        #if NUM_PICSTYLES > 9 // 600D, 5D3...
+                        "Auto",
+                        #endif
+                        "Standard", "Portrait", "Landscape",
+                        #if NUM_PICSTYLES == 11 // D8 and up
+                        "FineDetail",
+                        #endif
+                        "Neutral", "Faithful", "Monochrome", "UserDef1", "UserDef2", "UserDef3" },
+                .update     = picstyle_display_submenu,
+                .select     = picstyle_toggle,
+                .help = "Change current picture style.",
+                //~ .show_liveview = 1,
+                .edit_mode = EM_SHOW_LIVEVIEW,
+                .icon_type = IT_DICE,
+            },
+            {
+                .name = "Sharpness",
+                .update     = sharpness_display,
+                .select     = sharpness_toggle,
+                .help = "Adjust sharpness in current picture style.",
+                .edit_mode = EM_SHOW_LIVEVIEW,
+            },
+            {
+                .name = "Contrast",
+                .update     = contrast_display,
+                .select     = contrast_toggle,
+                .help = "Adjust contrast in current picture style.",
+                .edit_mode = EM_SHOW_LIVEVIEW,
+            },
+            {
+                .name = "Saturation",
+                .update     = saturation_display,
+                .select     = saturation_toggle,
+                .help = "Adjust saturation in current picture style.",
+                .edit_mode = EM_SHOW_LIVEVIEW,
+            },
+            {
+                .name = "Color Tone",
+                .update     = color_tone_display,
+                .select     = color_tone_toggle,
+                .help = "Adjust color tone in current picture style.",
+                .edit_mode = EM_SHOW_LIVEVIEW,
+            },
+    #ifdef FEATURE_REC_PICSTYLE
+            {
+                .name = "REC-PicStyle",
+                .priv = &picstyle_rec,
+                .max  = NUM_PICSTYLES,
+                .icon_type = IT_DICE_OFF,
+                .update     = picstyle_rec_sub_display,
+                .select     = picstyle_rec_sub_toggle,
+
+                .choices = (const char *[]) {"OFF",
+                #if NUM_PICSTYLES == 10 // 600D, 5D3...
+                "Auto",
+                #endif
+                "Standard", "Portrait", "Landscape",
+                #if NUM_PICSTYLES == 11 // D8 and up
+                "FineDetail",
+                #endif
+                "Neutral", "Faithful", "Monochrome", "UserDef1", "UserDef2", "UserDef3" },
+                .help = "You can use a different picture style when recording.",
+                .depends_on = DEP_MOVIE_MODE,
+            },
+    #endif // FEATURE_REC_PICSTYLE
+            MENU_EOL
+        },
+    },
+};
+
+static struct lvinfo_item info_item = {
+    .name = "Pic.Style",
+    .which_bar = LV_TOP_BAR_ONLY,
+    .update = picstyle_update,
+    .priority = -1,
+    .preferred_position = 5,
+};
+
+
+static void picstyle_features_init()
+{
+    menu_add("Shoot", picstyle_features_menu, COUNT(picstyle_features_menu));
+    lvinfo_add_item(&info_item);
+}
+
+INIT_FUNC(__FILE__, picstyle_features_init);
+#endif // FEATURE_PICSTYLE
